requirements.txt
pandas
numpy
scikit-learn
SQLAlchemy
psycopg2-binary
python-dotenv